requirements.txt
# === Core AI and LangGraph ===
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=2.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-pinecone>=0.0.3

# === LLM and Chat Interfaces ===
openai>=1.7.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# === Data Handling and Processing ===
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.0.0
xlrd>=2.0.0
sqlalchemy>=2.0.0

# === Snowflake and DB Connectivity ===
snowflake-connector-python>=3.6.0

# === Embeddings and Vector DBs ===
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pinecone-client>=3.0.0
tiktoken>=0.5.2

# === RAG Document Processing ===
pypdf>=3.17.0
PyPDF2>=3.0.1
python-docx>=1.1.0
python-multipart>=0.0.6

# === Utilities ===
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
typing-extensions>=4.9.0
aiofiles>=23.2.1

# === Visualization ===
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# === Development & Testing ===
pytest>=7.4.0
black>=23.10.0
isort>=5.12.0


# === Kafka ===
confluent-kafka==2.3.0
snowflake-connector-python==3.6.0
requests==2.31.0
python-dotenv==1.0.0
